Return existing class from generated_classes
This is from `create_entity()`.

If an existing class already exists in the `generated_classes` module,
it is investigated whether the `entity` attribute matches the current
entity, if yes, the class is returned, if no to any of those conditions
a ValueError is raised.

In the future, one could either extend the naming conventions of these
classes to include more and more "parts" of the identity in order to get
uniqueness, supporting having entity instance classes for the same
entity with different versions.